+++ renet2/src/channel/mod.rs
@@ -11,7 +11,8 @@ pub(crate) use slice_constructor::SliceConstructor;
 pub enum SendType {
     // Messages can be lost or received out of order.
     Unreliable {
-        // Should only be `true` if they underlying transport is ordered and reliable (e.g. WebSockets or in-memory channels).
+        // Should only be `true` if the underlying transport is ordered and reliable (e.g. WebSockets or in-memory channels)
+        // and reliability is desired. Can be set to `false` if reliability is not required.`
         ordered_reliable_substrate: bool,
     },
     /// Messages are guaranteed to be received and in the same order they were sent.